refactor: rename profile helper files
